+++ common/src/main/java/depchain/common/domain/EOAccount.java
@@ -5,7 +5,7 @@ import java.security.KeyPair;
 public class EOAccount extends Account {
     private final KeyPair keyPair;
 
-    public EOAccount(String address, Double balance, KeyPair keyPair) {
+    public EOAccount(String address, long balance, KeyPair keyPair) {
         super(address, balance);
         this.keyPair = keyPair;
     }
